+++ changelog.md
@@ -5,7 +5,9 @@
 ### WIP
 
 - [add] 增加了 `ZhihuClient.from_url` 方法，传入合法的知乎网址，就能生成对应的对象
-- [add] 给`BaseGenerator`增加了 `add_params`和`set_params` 方法
+- [add] 给 `BaseGenerator` 增加了 `add_params` 和 `set_params`  方法
+- [fix] 修复了 `BaseGenerator` 在 Python 2 下有问题的情况。
+- [fix] 修复了当用户的 `locations`，`educations`，`business`，`employments` 等属性值不存在强行获取会出错的 bug
 - [add] 写完了文档
 - [change] 改变了好多内部类名和变量名，不过对外部接口没有影响